MC-15: Refactor components into one dir per component
